+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/ExpectedFormHierarchies.swift
@@ -265,9 +265,6 @@ enum ExpectedFormHierarchy {
                 FormHierarchyNode(type: "SectionElement", children: [
                     FormHierarchyNode(type: "TextFieldElement", properties: ["label": "Full name"])
                 ]),
-                FormHierarchyNode(type: "SectionElement", children: [
-                    FormHierarchyNode(type: "DropdownFieldElement", properties: ["itemCount": "14", "label": "iDEAL Bank"])
-                ]),
             ])
         }
 
@@ -276,9 +273,6 @@ enum ExpectedFormHierarchy {
                 FormHierarchyNode(type: "SectionElement", properties: ["title": "Contact information"], children: [
                     FormHierarchyNode(type: "TextFieldElement", properties: ["label": "Full name"]),
                     FormHierarchyNode(type: "TextFieldElement", properties: ["label": "Email"]),
-                ]),
-                FormHierarchyNode(type: "SectionElement", children: [
-                    FormHierarchyNode(type: "DropdownFieldElement", properties: ["itemCount": "14", "label": "iDEAL Bank"])
                 ]),
                 FormHierarchyNode(type: "SimpleMandateElement", properties: ["text": "By providing your payment information and confirmi..."]),
             ])